LPS-104294 Move DummyPortalCache out into internal package.
